Added more functionality to ObjectDescription. Implemented IDescribable-type functions on ConstantNode, for testing. Fixed TypeName bug. Fixed a couple of int->double warnings.

+++ libraries/features/src/MeanFeature.cpp
@@ -97,7 +97,7 @@ namespace features
         }
 
         auto sumCoordinates = model.EmplaceLayer<layers::Sum>(perChannelBufferOutputCoordinates);
-        auto divisorCoordinates = model.EmplaceLayer<layers::ConstantLayer>(std::vector<double>(dimension, windowSize));
+        auto divisorCoordinates = model.EmplaceLayer<layers::ConstantLayer>(std::vector<double>(dimension, static_cast<double>(windowSize)));
         
         auto meanCoordinates = model.EmplaceLayer<layers::BinaryOperationLayer>(sumCoordinates, divisorCoordinates, layers::BinaryOperationLayer::OperationType::divide);
         return meanCoordinates;
